move ensure_tz,ListOptions,PostOptions,PostOptions to single shared common.py
