+++ python/svix/api/message_attempt.py
@@ -1,8 +1,10 @@
 import typing as t
-from dataclasses import asdict, dataclass
-from datetime import datetime, timezone
+from dataclasses import dataclass
+from datetime import datetime
 
 from deprecated import deprecated
+
+from .common import ensure_tz, ListOptions, PostOptions, ApiBase
 
 
 from ..internal.openapi_client.api.message_attempt import (
@@ -15,7 +17,6 @@ from ..internal.openapi_client.api.message_attempt import (
     v1_message_attempt_list_by_msg,
     v1_message_attempt_resend,
 )
-from ..internal.openapi_client.client import AuthenticatedClient
 from ..internal.openapi_client.models.list_response_endpoint_message_out import (
     ListResponseEndpointMessageOut,
 )
@@ -35,32 +36,6 @@ from ..internal.openapi_client.models.status_code_class import StatusCodeClass
 DEFAULT_SERVER_URL = "https://api.svix.com"
 
 
-def ensure_tz(x: t.Optional[datetime]) -> t.Optional[datetime]:
-    if x is None:
-        return None
-
-    if x.tzinfo is None:
-        return x.replace(tzinfo=timezone.utc)
-    return x
-
-
-@dataclass
-class ListOptions:
-    iterator: t.Optional[str] = None
-    limit: t.Optional[int] = None
-
-    def to_dict(self) -> t.Dict[str, t.Any]:
-        return {k: v for k, v in asdict(self).items() if v is not None}
-
-
-@dataclass
-class PostOptions:
-    idempotency_key: t.Optional[str] = None
-
-    def to_dict(self) -> t.Dict[str, t.Any]:
-        return {k: v for k, v in asdict(self).items() if v is not None}
-
-
 @dataclass
 class MessageAttemptListOptions(ListOptions):
     status: t.Optional[MessageStatus] = None
@@ -77,13 +52,6 @@ class MessageAttemptListOptions(ListOptions):
         if self.after is not None:
             d["after"] = ensure_tz(self.after)
         return d
-
-
-class ApiBase:
-    _client: AuthenticatedClient
-
-    def __init__(self, client: AuthenticatedClient) -> None:
-        self._client = client
 
 
 class MessageAttemptAsync(ApiBase):
