+++ python/svix/api/statistics.py
@@ -1,35 +1,18 @@
-import typing as t
-from dataclasses import asdict, dataclass
-
-
-from svix.internal.openapi_client.models.aggregate_event_types_out import (
+from ..internal.openapi_client.models.aggregate_event_types_out import (
     AggregateEventTypesOut,
 )
-from svix.internal.openapi_client.models.app_usage_stats_in import AppUsageStatsIn
-from svix.internal.openapi_client.models.app_usage_stats_out import AppUsageStatsOut
+from ..internal.openapi_client.models.app_usage_stats_in import AppUsageStatsIn
+from ..internal.openapi_client.models.app_usage_stats_out import AppUsageStatsOut
+
+from .common import PostOptions, ApiBase
+
 
 from ..internal.openapi_client.api.statistics import (
     v1_statistics_aggregate_app_stats,
     v1_statistics_aggregate_event_types,
 )
-from ..internal.openapi_client.client import AuthenticatedClient
 
 DEFAULT_SERVER_URL = "https://api.svix.com"
-
-
-@dataclass
-class PostOptions:
-    idempotency_key: t.Optional[str] = None
-
-    def to_dict(self) -> t.Dict[str, t.Any]:
-        return {k: v for k, v in asdict(self).items() if v is not None}
-
-
-class ApiBase:
-    _client: AuthenticatedClient
-
-    def __init__(self, client: AuthenticatedClient) -> None:
-        self._client = client
 
 
 class StatisticsAsync(ApiBase):
